+++ app/femr/ui/views/admin/trips/edit.scala.html
@@ -1,7 +1,7 @@
-@(currentUser: femr.common.dtos.CurrentUser, viewModel: femr.ui.models.superuser.trips.EditViewModelGet)
+@(currentUser: femr.common.dtos.CurrentUser, viewModel: femr.ui.models.admin.trips.EditViewModelGet)
 
 @import femr.ui.views.html.layouts.admin
-@import femr.ui.controllers.superuser.routes.TripController
+@import femr.ui.controllers.admin.routes.TripController
 
 @additionalStyles = {
