+++ app/femr/ui/views/partials/admin/trips/tripsMenu.scala.html
@@ -1,4 +1,4 @@
-@import femr.ui.controllers.superuser.routes.TripController
+@import femr.ui.controllers.admin.routes.TripController
 
 <div class="tripOptionsWrap">
     <a href="@TripController.manageGet()" class="fButton fOtherButton fAdminButton userBtns"><span class="glyphicon glyphicon-plus-sign"></span>
