+++ conf/routes
@@ -7,14 +7,6 @@ GET         /superuser/tabs/:name                              @femr.ui.controll
 POST        /superuser/tabs/:name                              @femr.ui.controllers.superuser.TabController.fieldsPost(name: String)
 GET         /superuser/tabs                                    @femr.ui.controllers.superuser.TabController.manageGet()
 POST        /superuser/tabs                                    @femr.ui.controllers.superuser.TabController.managePost()
-GET         /superuser/trips                                   @femr.ui.controllers.superuser.TripController.manageGet()
-POST        /superuser/trips                                   @femr.ui.controllers.superuser.TripController.managePost()
-GET         /superuser/trips/edit/:id                          @femr.ui.controllers.superuser.TripController.editGet(id: Integer)
-POST        /superuser/trips/edit/:id                          @femr.ui.controllers.superuser.TripController.editPost(id: Integer)
-GET         /superuser/trips/teams                             @femr.ui.controllers.superuser.TripController.teamsGet()
-POST        /superuser/trips/teams                             @femr.ui.controllers.superuser.TripController.teamsPost()
-GET         /superuser/trips/cities                            @femr.ui.controllers.superuser.TripController.citiesGet()
-POST        /superuser/trips/cities                            @femr.ui.controllers.superuser.TripController.citiesPost()
 GET         /superuser                                         @femr.ui.controllers.superuser.SuperuserController.indexGet()
 #Admin
 POST        /admin/users/create                                @femr.ui.controllers.admin.UsersController.createPost()
@@ -29,6 +21,14 @@ POST        /admin/inventory/get                               @femr.ui.controll
 POST        /admin/inventory/delete/:id                        @femr.ui.controllers.admin.InventoryController.ajaxDelete(id: Integer)
 GET         /admin/configure                                   @femr.ui.controllers.admin.ConfigureController.manageGet()
 POST        /admin/configure                                   @femr.ui.controllers.admin.ConfigureController.managePost()
+GET         /admin/trips                                       @femr.ui.controllers.admin.TripController.manageGet()
+POST        /admin/trips                                       @femr.ui.controllers.admin.TripController.managePost()
+GET         /admin/trips/edit/:id                              @femr.ui.controllers.admin.TripController.editGet(id: Integer)
+POST        /admin/trips/edit/:id                              @femr.ui.controllers.admin.TripController.editPost(id: Integer)
+GET         /admin/trips/teams                                 @femr.ui.controllers.admin.TripController.teamsGet()
+POST        /admin/trips/teams                                 @femr.ui.controllers.admin.TripController.teamsPost()
+GET         /admin/trips/cities                                @femr.ui.controllers.admin.TripController.citiesGet()
+POST        /admin/trips/cities                                @femr.ui.controllers.admin.TripController.citiesPost()
 GET         /admin                                             @femr.ui.controllers.admin.AdminController.index()
 #Pharmacy
 GET         /pharmacy/edit/:id                                 @femr.ui.controllers.PharmaciesController.editGet(id: Integer)
